% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LW1.R
\name{LW1}
\alias{LW1}
\title{LW1}
\usage{
LW1(y, alphas, betas, tau2s, xs, delta)
}
\arguments{
\item{y}{representa la serie de observaciones reales.}

\item{alphas}{representa la matriz de observaciones de contraste.}

\item{betas}{bla bla bla, Omar debe completar esto.}

\item{tau2s}{bla bla bla, Omar debe completar esto.}

\item{xs}{bla bla bla, Omar debe completar esto.}

\item{delta}{bla bla bla, Omar debe completar esto.}
}
\value{
Esta funcion retorna bla bla bla, Omar debe completar esto..
}
\description{
Función de filtro de partículas con aprendizaje de parámetros enfoque Liu & West
Basada en el algoritmo de Lopes (2014)
}
\examples{
# Example that Omar need to complete

#LW1(y=c(1, 2), alphas=c(2, 3), betas=c(1, 3),
#    tau2s=c(2, 4), xs=4, delta=1)

hist(rnorm(n=200))
}
\author{
Omar Rios Saavedra, \email{orioss@unal.edu.co}
}
