% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LWW.R
\name{LWW}
\alias{LWW}
\title{LWW}
\usage{
LWW(y, alphas, betas, tau2s, xs, delta, lev, M = 75, Ne = 20, method = 1)
}
\arguments{
\item{y}{representa la serie de observaciones reales.}

\item{alphas}{representa la matriz de observaciones de contraste.}

\item{betas}{bla bla bla, Omar debe completar esto.}

\item{tau2s}{bla bla bla, Omar debe completar esto.}

\item{xs}{bla bla bla, Omar debe completar esto.}

\item{delta}{bla bla bla, Omar debe completar esto.}

\item{lev}{bla bla bla, Omar debe completar esto.}

\item{M}{bla bla bla, Omar debe completar esto.}

\item{Ne}{bla bla bla, Omar debe completar esto.}

\item{method}{bla bla bla, Omar debe completar esto.}
}
\value{
Esta funcion retorna bla bla bla, Omar debe completar esto..
}
\description{
Esta funcion calcula el modelo de volatilidad estocástica basada en el filtro de partículas de Liu & West
el algoritmo incorpora los pasos de empuje bayesianos basados en la transformación waveltet.
plWav1j, metodología propuesta para la eliminación de ruido aditivo basado en la transformación wavelet.
BAYES.THR, metodología de Abramovich (1998) para la eliminación de ruido aditivo basado en la transformación wavelet.
}
\examples{
# Example that Omar need to complete

plot(density(rnorm(n=200)), lwd=4, col="tomato")
}
\author{
Omar Rios Saavedra, \email{orioss@unal.edu.co}
}
