% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plWav1j_copia.R
\name{plWav1j_copia}
\alias{plWav1j_copia}
\title{plWav1j_copia}
\usage{
plWav1j_copia(
  dat,
  filter.number = 4,
  family = "DaubLeAsymm",
  a = 10,
  b = 10,
  bet = 1,
  nu0 = 5,
  lamb0 = 10,
  M = 10,
  Ne,
  j0 = nlevelsWT(vw),
  plot.EMPL = FALSE
)
}
\arguments{
\item{dat}{bla bla bla, Omar debe completar esto.}

\item{filter.number}{bla bla bla, Omar debe completar esto.}

\item{family}{bla bla bla, Omar debe completar esto.}

\item{a}{bla bla bla, Omar debe completar esto.}

\item{b}{bla bla bla, Omar debe completar esto.}

\item{bet}{bla bla bla, Omar debe completar esto.}

\item{nu0}{bla bla bla, Omar debe completar esto.}

\item{lamb0}{bla bla bla, Omar debe completar esto.}

\item{M}{bla bla bla, Omar debe completar esto.}

\item{Ne}{bla bla bla, Omar debe completar esto.}

\item{j0}{bla bla bla, Omar debe completar esto.}

\item{plot.EMPL}{bla bla bla, Omar debe completar esto.}
}
\value{
Esta funcion retorna bla bla bla, Omar debe completar esto..
}
\description{
Esta funcion calcula bla bla bla.
}
\examples{
# Example that Omar need to complete

hist(rpois(n=200, lambda=10), col="pink")
}
\author{
Omar Rios Saavedra, \email{orioss@unal.edu.co}
}
