% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LW1.R
\name{LW1}
\alias{LW1}
\title{LW1}
\usage{
LW1(y, alphas, betas, tau2s, xs, delta)
}
\arguments{
\item{y}{representa la serie de observaciones reales.}

\item{alphas}{representa los valores iniciales para el parámetro de reversión de la media en el proceso de volatilidad estocástica.}

\item{betas}{representa los valores iniciales para el parámetros de persistencia de volatilidad.}

\item{tau2s}{representa los valorese iniciales para la varianza de la variable latente (volatilidad estocástica).}

\item{xs}{partículas iniciales de la variable latente a partir de la distribución a priori.}

\item{delta}{constante de ponderación para el aprendizaje de parámetros en el algoritmo Liu & West (2001).}
}
\value{
Esta funcion retorna los cuantiles (2.5\%, 50\% y 97.5\%) de las estimaciones de la volatilidad estocástica,
y sus parámetros (\eqn{\alpha}, \eqn{\beta} y \eqn{\tau^2}).
}
\description{
Función de filtro de partículas con aprendizaje de parámetros enfoque Liu & West
Basada en el algoritmo de Lopes & Tsay (2011)
}
\examples{
# Example that Omar need to complete

#LW1(y=c(1, 2), alphas=c(2, 3), betas=c(1, 3),
#    tau2s=c(2, 4), xs=4, delta=1)

hist(rnorm(n=200))
}
\author{
Omar Rios Saavedra, \email{orioss@unal.edu.co}
}
